Refatorando os os Controllers e Services de User e Post

+++ src/main/java/com/example/blog/services/CommentService.java
@@ -1,14 +1,18 @@
 package com.example.blog.services;
 
 import com.example.blog.entities.comments.Comment;
+import com.example.blog.entities.comments.dtos.AddCommentDTO;
 import com.example.blog.entities.post.Post;
 import com.example.blog.entities.user.User;
+import com.example.blog.entities.user.UserRole;
+import com.example.blog.exceptions.EntityNotFoundException;
 import com.example.blog.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CommentService {
@@ -22,7 +26,13 @@ public class CommentService {
     @Autowired
     CommentRepository commentRepository;
 
-    public void create(Comment comment, Post post, User user){
+    public void create(AddCommentDTO data, Post post, User user){
+
+
+
+        LocalDate now = LocalDate.now();
+        Comment comment = new Comment(null, data.content(), user, post, now);
+
         commentRepository.save(comment);
         post.getComments().add(comment);
         postService.save(post);
@@ -32,11 +42,10 @@ public class CommentService {
 
     public void save(Comment comment){ commentRepository.save(comment); }
 
-    public List<Comment> findAll(){
-        return commentRepository.findAll();
+    public Comment findById(String id) throws EntityNotFoundException {
+        return  commentRepository.findById(id).
+            orElseThrow(() -> new EntityNotFoundException("Comment not found"));
     }
-
-    public Optional<Comment> findById(String id) { return  commentRepository.findById(id);}
 
     public void likeOrDislike(Comment comment, User user){
 
@@ -53,22 +62,29 @@ public class CommentService {
 
     }
 
-    public void delete(Comment comment) {
+    public void delete(Comment comment, User user) {
 
-        comment.getAuthor().getComments().remove(comment);
-        userService.save(comment.getAuthor());
+        if (comment.getAuthor().equals(user) || user.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals(UserRole.ADMIN.name()))){
 
-        comment.getPost().getComments().remove(comment);
-        postService.save(comment.getPost());
+            comment.getAuthor().getComments().remove(comment);
+            userService.save(comment.getAuthor());
 
-        List<User> likesOnComment = comment.getLikes();
-        for(User user : likesOnComment){
-            user.getLikedComments().remove(comment);
-            userService.save(user);
+            comment.getPost().getComments().remove(comment);
+            postService.save(comment.getPost());
+
+            List<User> likesOnComment = comment.getLikes();
+            for(User user_who_likes : likesOnComment){
+                user_who_likes.getLikedComments().remove(comment);
+                userService.save(user_who_likes);
+            }
+
+
+            commentRepository.delete(comment);
+        }
+        else {
+            throw new AccessDeniedException("You must be owner of comment");
         }
 
-
-        commentRepository.delete(comment);
     }
 
 }
