+++ src/main/java/com/example/blog/services/PostService.java
@@ -2,13 +2,17 @@ package com.example.blog.services;
 
 import com.example.blog.entities.comments.Comment;
 import com.example.blog.entities.post.Post;
+import com.example.blog.entities.post.dtos.CreatePostDTO;
 import com.example.blog.entities.user.User;
+import com.example.blog.entities.user.UserRole;
 import com.example.blog.entities.user.dtos.UsernameDTO;
+import com.example.blog.exceptions.EntityNotFoundException;
 import com.example.blog.repositories.CommentRepository;
 import com.example.blog.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,7 +28,11 @@ public class PostService {
     @Autowired
     CommentRepository commentRepository;
 
-    public void create(Post post, User user){
+    public void create(CreatePostDTO data, User user){
+
+        LocalDate now = LocalDate.now();
+        Post post = new Post(data.title(), data.content(), user, now);
+
         postRepository.save(post);
         user.getPosts().add(post);
         userService.save(user);
@@ -34,7 +42,11 @@ public class PostService {
         return postRepository.findAll();
     }
 
-    public Optional<Post> findById(String id) { return postRepository.findById(id); }
+    public Post findById(String id) throws EntityNotFoundException {
+        return postRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Post not found"));
+    }
+
 
     public void save(Post post) {  postRepository.save(post); }
 
@@ -53,29 +65,34 @@ public class PostService {
 
     }
 
-    public void delete(Post post){
+    public void delete(Post post, User whoRequestDelete) {
 
-        post.getAuthor().getPosts().remove(post);
-        userService.save(post.getAuthor());
+        if (post.getAuthor().equals(whoRequestDelete) || whoRequestDelete.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals(UserRole.ADMIN.name()))) {
 
-        List<Comment> comments = post.getComments();
-        for (Comment comment : comments) {
-            List<User> likedUsers = comment.getLikes();
+
+            post.getAuthor().getPosts().remove(post);
+            userService.save(post.getAuthor());
+
+            List<Comment> comments = post.getComments();
+            for (Comment comment : comments) {
+                List<User> likedUsers = comment.getLikes();
+                for (User user : likedUsers) {
+                    user.getLikedComments().remove(comment);
+                    userService.save(user);
+                }
+                commentRepository.delete(comment);
+            }
+
+            List<User> likedUsers = post.getLikes();
             for (User user : likedUsers) {
-                user.getLikedComments().remove(comment);
+                user.getLikedPosts().remove(post);
                 userService.save(user);
             }
-            commentRepository.delete(comment);
+
+
+            postRepository.delete(post);
+        } else {
+            throw new SecurityException("You do not have permission to delete this post");
         }
-
-        List<User> likedUsers = post.getLikes();
-        for (User user : likedUsers) {
-            user.getLikedPosts().remove(post);
-            userService.save(user);
-        }
-
-
-
-        postRepository.delete(post);
     }
 }
